requirements.txt
scipy==1.2.1
numpy==1.19.1
matplotlib-base==3.3.1
pillow==7.2.0
pandas==1.1.1
gdal==2.4.2
requests==2.24.0
json-c==0.13.1